fix: error for large sample sizes in Multinomial NB
I was seeing SystemStackError issues with Multinomial NB when running
with larger sample sizes. I believe this is due to splatting the
samples into `Numo::DFloat` as params. That param list can be very
large with large sample sets, which overflows Ruby's default heap
size.

I think instead we can use `Numo::DFloat.cast` here, passing in the
array, as that will just be a single param.

I've also lifted this cast to the `bin_x` variable so we don't have to
do it per class.